Redirect requrests to canonical host
This may be better handled in the reverse proxy. It is difficult to know
where we are currently running. Is this a development server? Can the
development server handle HTTPS? What if we want to run the development
server in release mode? Should we duplicate the canonicalize function
across multiple services? What if someone else is hosting files? This
may become overly complicated.